Add UrlSanitizerTest
Move the url.full redaction cases out of HttpClientAttributesExtractorTest,
which built a full extractor and getter just to assert a string transformation,
and add coverage for edge cases that were only reachable indirectly: null and
empty input, URLs with no scheme or no authority, and an empty sensitive
parameter set (where query redaction early-returns but userinfo redaction must
still apply).

Copilot-Session: a673f464-c6f9-4dc2-b840-227adecbe243
